% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedHypoRate.R
\name{ExpectedHypoRate}
\alias{ExpectedHypoRate}
\title{Calculating the expected (true) hypothesis rate(s).}
\usage{
ExpectedHypoRate(
  ES,
  VCOV,
  NrHypos,
  Hypos,
  Safeguard = "unconstrained",
  nsim = 1000,
  seed = 123
)
}
\arguments{
\item{VCOV}{A square (p times p) matrix denoting the covariance matrix of the (standardized) parameters. It is assumed that the (standardized) parameter estimates come from a multivariate normal distribution with covariance matrix VCOV.}

\item{NrHypos}{The number of theory-based hypotheses in the set of candidate hypotheses (is a scalar with an integer value).}

\item{Hypos}{A vector of strings containing the NrHypos theory-based hypotheses (on the p estimates). These should either use the names X1 to Xp or the names (characters) given to mu (see the last example below).}

\item{Safeguard}{Indicator of which safeguard-hypothesis should be used: "unconstrained" (default; i.e., all possible theories including the one specified), "none" (only advised when set of hyptheses cover all theories), or (only when 'NrHypos = 1') "complement" (i.e., the remaining theories).}

\item{nsim}{Optional. The number of iterations used in the calculation. By default, nsim = 1000. The higher nsim, the higher the precision but als the higher the computation time.}

\item{seed}{Optional. The seed value used in the calculation. By default, seed = 123. This can be used for a sensitivty check. If for another seed value the results unnegligibly differ, then nsim has to be increased.}

\item{mu}{A (named) vector (of length p) denoting the population value of the (standardized) parameters. It is assumed that the (standardized) parameter estimates come from a multivariate normal distribution with mean mu.}
}
\value{
This function renders 3 types of output.
The first (mle_Hypos) is the percentage of times the sampled (maximum likelihood) estimates are in agreement with the hypotheses in the set, when sampled from a multivariate normal distribution with mean mu and covariance matrix VCOV. This is independent from the GORICA, in a way purely based on the log likelhood. Bear in mind that the estimates are always in agreement with the unconstrained hypothesis.
The second (HR_Hypos) is the percentage of times each hypothesis in the set is preferred under that distribution; rendering the Hypothesis Rate (HR) for H1. This is based on the GORICA and thus both the log likelihood and the penalty are taken into account.
The third are the average GORICA weight values for the hypotheses in the set; thus either for H1 and its complement or H1 and the unconstrained hypothesis.
}
\description{
This function calculates the expected (true) hypothesis rate for the hypotheses in the set under a specific hypothesis (say, H0 or H1 or even another hypothesis).
}
\examples{

# Example based on 6 standardized parameters of interest.
p <- 6
mu <- rep(0, p) # mu is in accordance with H0: X1 = X2 = X3 = X4 = X5 = X6 (or H0 <- "X1 == X2; X3 == X4; X5 == X6")
VCOV <- diag(p) # Here, the variances of the parameters are 1 and the covariances are 0
NrHypos <- 1
H1 <- "X1 < X2; X3 < X4; X5 < X6"
Hypos <- c(H1)
Safeguard <- "complement"
ExpectedHypoRate(mu, VCOV, NrHypos, Hypos, Safeguard)

# Other covariance matrix of the standardized parameters #
# If we assume a correlation between the standardized parameters of rho = 0.25 and a sample size of 100, then
rho <- 0.25
n <- 100
sigma <- matrix(rho, ncol = p, nrow = p)
for(i in 1:p){
 sigma[i,i] <- 1
}
VCOV <- sigma / n
ExpectedHypoRate(mu, VCOV, NrHypos, Hypos, Safeguard)

# Other mean vector of the standardized parameters #
# In the example above, it is assumed that the truth is a null hypothesis (mu = vector of 0's).
# One can also assume another distribution, e.g. stating that the standardized estimates have a specific effect:
mu <- c(-0.1, 0.1, -0.1, 0.1, -0.1, 0.1) # mu is now in accordance with H1 <- "X1 < X2; X3 < X4; X5 < X6"
# One can do this for multiple values corresponding to specific effect sizes for example.
ExpectedHypoRate(mu, VCOV, NrHypos, Hypos, Safeguard)

# Multiple hypotheses of interest #
NrHypos <- 2
H1 <- "X1 < X2; X3 < X4; X5 < X6"
H2 <- "X1 > X2; X3 > X4; X5 > X6"
Hypos <- c(H1, H2) # Currently, in case of multiple theory-based hypotheses in the set, one cannot use the complement of these as safeguard, so:
Safeguard <- "unconstrained" # which is the default, so could be left out
ExpectedHypoRate(mu, VCOV, NrHypos, Hypos, Safeguard)

# Example with named vector mu.
p <- 6
mu <- rep(0, p) # mu is in accordance with H0: X1 = X2 = X3 = X4 = X5 = X6 (or H0 <- "X1 == X2; X3 == X4; X5 == X6")
names(mu) <- c("beta12", "beta21", "beta13", "beta31", "beta23", "beta32")
VCOV <- diag(p) # Here, the variances of the parameters are 1 and the covariances are 0
NrHypos <- 1
H1 <- "beta12 < beta21; beta13 < beta31; beta23 < beta32" # This should now use the names of mu
Hypos <- c(H1)
Safeguard <- "complement"
ExpectedHypoRate(mu, VCOV, NrHypos, Hypos, Safeguard)

}
