% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myIC.R
\docType{data}
\name{myIC}
\alias{myIC}
\title{An example of IC values}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
data(myIC)
}
\description{
An example of IC: A vector with 3 information criterion values
}
\examples{
data(myIC)

}
\keyword{datasets}
