% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICweights.r
\name{IC.weights}
\alias{IC.weights}
\title{Calculating IC weights based on IC values (AIC, ORIC, GORIC(A), BIC, SIC, ...)}
\usage{
IC.weights(IC, Name_Hypo = NULL)
}
\arguments{
\item{IC}{A vector or one-column matrix with information criteria (AIC, ORIC, GORIC(A), BIC, SIC, ...) values of length 'NrHypos', where 'NrHypos' stands for the number of hypotheses/models.}

\item{Name_Hypo}{Optional. Vector containing 'NrHypos' characters which will be used for labeling the hypothesis. Default: H1, H2, ....}
}
\value{
IC weights, which quantify the relative strength of hypotheses/models.
}
\description{
This function transforms IC values into IC weights: IC values denote the ordering of hypotheses/models, while IC weights quantify the relative strength of hypotheses/models.
}
\examples{

# library(ICweights)

IC <- myIC # Example based on 3 hypotheses.
IC.weights(IC)

# Change labels of hypotheses #
#
# Example 1: Let us say that we evaluated H1: beta1 > beta2; H2: beta1 = beta2; and H3: beta1 < beta2.
Name_Hypo <- c("Higher", "Equal", "Smaller")
IC.weights(IC, Name_Hypo)
#
# Example 2: Let us say that we evaluated a linear, quadratic, and cubic model.
# Notably, this can also be represented by
# H1: beta_linear, beta_quadratic = 0, beta_cubic = 0;
# H2: beta_linear, beta_quadratic, beta_cubic = 0; and
# Hunc: beta_linear, beta_quadratic, beta_cubic (i.e., no restrictions on the beta's).
Name_Hypo <- c("Linear", "Quadratic", "Cubic")
IC.weights(IC, Name_Hypo)

# Example 3: This examples shows some output options
ICw <- IC.weights(IC)
print(ICw)
summary(ICw)
print(ICw, digits = 4)
summary(ICw, digits = 4)
# In Rstudio, use 'ICw$' to see what output there is. The following is available:
#' ICw$IC
ICw$IC.weights
ICw$ratio.IC.weights

}
