% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICweights.r
\name{IC.weights}
\alias{IC.weights}
\title{Calculating IC weights based on IC values (AIC, ORIC, GORIC(A), BIC, SIC, ...)}
\usage{
IC.weights(IC, Name_Hypo = NULL)
}
\arguments{
\item{IC}{A vector or one-column matrix with information criteria (AIC, ORIC, GORIC(A), BIC, SIC, ...) values of length 'NrHypos', where 'NrHypos' stands for the number of hypotheses/models.}

\item{Name_Hypo}{Optional. Vector containing 'NrHypos' characters which will be used for labelling the hypothesis. Default: H1, H2, ....}
}
\value{
IC weights, which quantify the relative strength of hypotheses/models.
}
\description{
This function transforms IC values into IC weights: IC values denote the ordering of hypotheses/models, while IC weights quantify the relative strength of hypotheses/models.
}
\examples{

IC <- myIC # Example based on 3 hypotheses.
IC.weights(IC)

# Change labels of hypotheses #
# For example, let us say that we tested a linear model vs quadratic vs cubic.
Name_Hypo <- c("Linear", "Quadratic", "Cubic")
IC.weights(IC, Name_Hypo)
}
